@@ -1,6 +1,6 @@
 {
-  "name": "material-dashboard-angular",
-  "version": "2.1.0",
+  "name": "garrus-alpha",
+  "version": "0.1.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
@@ -12,19 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.2",
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "5.2.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "5.2.4",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/platform-server": "5.2.9",
-    "@angular/router": "5.2.9",
+    "@angular/animations": "6.0.9",
+    "@angular/cdk": "^6.4.1",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "6.0.9",
+    "@angular/core": "6.0.9",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/material": "^6.4.1",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/platform-server": "6.0.9",
+    "@angular/router": "6.0.9",
     "ajv": "6.4.0",
     "arrive": "2.4.1",
     "bootstrap": "4.1.0",
@@ -37,21 +36,24 @@
     "hammerjs": "2.0.8",
     "jquery": "3.2.1",
     "moment": "2.18.1",
-    "ngx-toastr": "^8.9.0",
-    "npm": "^5.6.0",
+    "ngx-toastr": "^8.10.0",
+    "node-sass": "^4.9.2",
+    "npm": "^5.10.0",
     "perfect-scrollbar": "1.1.0",
     "popper.js": "1.14.3",
-    "rxjs": "5.5.10",
+    "rxjs": "6.2.2",
+    "rxjs-compat": "^6.2.2",
     "web-animations-js": "2.3.1",
-    "zone.js": "0.8.4"
+    "zone.js": "0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
+    "@angular-devkit/build-angular": "^0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "6.0.9",
+    "@angular/language-service": "6.0.9",
     "@types/bootstrap": "3.3.32",
     "@types/chartist": "0.9.34",
-    "@types/googlemaps": "^3.30.8",
+    "@types/googlemaps": "^3.30.11",
     "@types/jasmine": "2.5.38",
     "@types/jquery": "1.10.31",
     "@types/node": "6.0.73",
@@ -67,6 +69,6 @@
     "protractor": "5.3.1",
     "ts-node": "5.0.1",
     "tslint": "5.9.1",
-    "typescript": "2.4.2"
+    "typescript": "2.7.2"
   }
 }
